+++ examples/tpch/queries/Q10.h
@@ -1,5 +1,5 @@
 // c_custkey, c_name, revenue, c_acctbal, n_name, c_address, c_phone, c_comment
-using Q10_relt = tuple<dkey_t, char*, float, float, char*, char*, char*, char*>;
+using Q10_relt = tuple<dkey_t, char*, double, double, char*, char*, char*, char*>;
 using Q10_rtype = sequence<Q10_relt>;
 
 Q10_rtype Q10(maps m, const char* start, const char* end) {
@@ -7,23 +7,23 @@ Q10_rtype Q10(maps m, const char* start, const char* end) {
   // take range of order dates
   o_order_map order_range = o_order_map::range(m.orders_for_date, Date(start), Date(end));
 	
-  using kf_pair = pair<dkey_t, float>;
+  using kf_pair = pair<dkey_t, double>;
 
   // get (custkey,revenue) for lineitems with return flag by order
   auto date_f = [&] (order_map::E& e) -> kf_pair {
     Orders& o = e.second.first;
-    auto li_f = [&] (line_item_set::E& l) -> float {
+    auto li_f = [&] (line_item_set::E& l) -> double {
       if (l.returnflag() != 'R') return 0.0;
       else return l.e_price*(1 - l.discount.val());
     };
-    float v = line_item_set::map_reduce(e.second.second, li_f, Add<float>());
+    double v = line_item_set::map_reduce(e.second.second, li_f, Add<double>());
     return make_pair(o.customer_key, v);
   };
   sequence<kf_pair> elts = flatten<kf_pair>(order_range, date_f);
 
   // sum revenue by custkey
-  sequence<kf_pair> r = parlay::reduce_by_key(elts, parlay::plus<float>());
-  //sequence<kf_pair> r = parlay::internal::group_by_and_combine(elts, parlay::addm<float>());
+  sequence<kf_pair> r = parlay::reduce_by_key(elts, parlay::plus<double>());
+  //sequence<kf_pair> r = parlay::internal::group_by_and_combine(elts, parlay::addm<double>());
 
   // sort by revenue
   auto less = [] (kf_pair a, kf_pair b) {return a.second > b.second;};
@@ -32,7 +32,7 @@ Q10_rtype Q10(maps m, const char* start, const char* end) {
   // generate results
   auto get_result = [&] (size_t i) -> Q10_relt {
     dkey_t custkey = res[i].first;
-    float revenue = res[i].second;
+    double revenue = res[i].second;
     Customer c = (*(m.orders_for_customer.find(custkey))).first;
     char* n_name = nations[c.nationkey].name;
     return Q10_relt(custkey, c.name(), revenue, c.acctbal,
